Store sections as ints, make the Section class use the Mnemonic class.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1174 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Cache.java
@@ -552,7 +552,7 @@ verifyRecords(Cache tcache) {
 }
 
 private final byte
-getCred(short section, boolean isAuth) {
+getCred(int section, boolean isAuth) {
 	if (section == Section.ANSWER) {
 		if (isAuth)
 			return Credibility.AUTH_ANSWER;
